inference/requirements.txt
fastapi
python-multipart
uvicorn
asyncpg
pydantic_settings
aiokafka
aiomultiprocess
opencv-python
Pillow
imageio_ffmpeg
ultralytics
imageai
 #pip install -r requirements.txt